fix stream type

+++ src/main/java/com/north6960/subsystems/vision/Limelight.java
@@ -67,7 +67,7 @@ public class Limelight {
     }
 
     public static void setCameraStream(StreamType streamType) {
-        table.getEntry("stream").setNumber(streamType.ordinal());
+        table.getEntry("stream").setNumber((double) streamType.ordinal()); 
     }
 
     public static double getCameraStream() {
